requirements.txt
colormath==3.0.0
hilbertcurve==2.0.5
numpy==2.1.1
Pillow==10.4.0
